FastAPIBackend/requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
feedparser
transformers
torch
sentencepiece
pydantic
pydantic-settings
python-dotenv
python-multipart
sqlmodel
sqlalchemy
psycopg2-binary
ta
joblib
scikit-learn
xgboost
cachetools
